Fix initWaitForSyncingFalse not working when syncing
The socket wasn't being open when it was syncing causing the web
app not to work, clients should be connected at all times and the
ux will show the node as syncing.

Another issue is that the syncing didn't return the correct type
when it is valid. It isn't true, it should be an obj